fix: save credential when edit a data connector
